+++ sdk/include/mods/svc/save.h
@@ -10,31 +10,20 @@
 #define SAVE_SERVICE_MAJOR 1u
 #define SAVE_SERVICE_MINOR 0u
 
-/* Handle for a save-observer registration. 0 is never a valid handle. */
+/* 0 is never a valid handle. */
 typedef uint64_t SaveObserverHandle;
 
-/* Total blob storage per mod per save slot. */
+/* Maximum combined blob size per mod and save slot. */
 #define SAVE_BLOB_BUDGET_BYTES 65536u
 
 /*
- * Per-save-slot mod data.
+ * Per-slot mod storage.
  *
- * Named binary blobs that travel with a save slot: stored in a host-owned sidecar next to the
- * emulated memory card, written when the game writes the card (in-game save, autosave), and
- * following file-select copy and erase. Blobs are namespaced per mod — two mods can use the
- * same name without collision. Format versioning inside a blob is the mod's own concern.
+ * Blobs are scoped to the calling mod and saved alongside each slot. Current-slot calls return
+ * MOD_UNAVAILABLE when no slot is active.
  *
- * Blob calls target the *current* slot, which exists from save load / new-save creation until
- * the player returns to file select; MOD_UNAVAILABLE otherwise. Data written between game
- * saves lives in memory only — like vanilla save state, it persists when the player saves.
- * A new save clears the slot's previous blob data before on_new_save fires.
- *
- * Staleness: the sidecar snapshots a checksum of each slot's vanilla save data at write time
- * and warns at load when they diverge (e.g. the card file was replaced externally); blobs are
- * still delivered — treat them with suspicion in that session.
- *
- * Callbacks run on the game thread. Registrations are owned by the calling mod and removed
- * automatically when it is disabled, reloaded, or fails.
+ * Callbacks run on the game thread. Observer registrations are removed when the calling mod is
+ * detached.
  */
 
 /* slot is the save-file index (0..2). */
@@ -43,46 +32,32 @@ typedef void (*SaveEventFn)(ModContext* ctx, uint32_t slot, void* user_data);
 typedef struct SaveService {
     ServiceHeader header;
 
-    /*
-     * Write a named blob for the current slot (copied). Fails with MOD_UNAVAILABLE when no
-     * slot is active or the mod's SAVE_BLOB_BUDGET_BYTES for the slot would be exceeded.
-     */
+    /* Store a copy in the current slot. Returns MOD_UNAVAILABLE if the limit would be exceeded. */
     ModResult (*set_blob)(ModContext* ctx, const char* name, const void* data, size_t size);
 
     /*
-     * Read a named blob from the current slot. With buf NULL, writes the blob's size to
-     * *inout_size. Otherwise *inout_size is the buffer capacity in, bytes written out.
-     * MOD_UNAVAILABLE when no slot is active or the blob does not exist.
+     * Read a blob from the current slot. Pass NULL for buf to query its size. Otherwise,
+     * inout_size is the buffer capacity on input and the blob size on success. Returns
+     * MOD_UNAVAILABLE if the blob does not exist.
      */
     ModResult (*get_blob)(ModContext* ctx, const char* name, void* buf, size_t* inout_size);
 
-    /* Remove a named blob from the current slot. */
     ModResult (*delete_blob)(ModContext* ctx, const char* name);
 
     /*
-     * Observe save lifecycle. Any callback may be NULL:
-     *  - on_new_save: a new file was finalized in file select (name entry complete); the
-     *    slot's blob store is empty — write initial blobs here. Like vanilla save data,
-     *    nothing hits disk until the first in-game save.
-     *  - on_save_loaded: a slot was loaded into the live game (file select or quick-load);
-     *    blobs are readable.
-     *  - on_save_written: the game persisted the slot (menu save or autosave); blob data
-     *    captured with it is on disk.
+     * Register save lifecycle callbacks. At least one callback is required. on_new_save runs
+     * after clearing the slot's blobs, on_save_loaded after activating the slot, and
+     * on_save_written after a successful game save. out_handle may be NULL.
      */
-    ModResult (*observe_saves)(ModContext* ctx, SaveEventFn on_new_save,
-        SaveEventFn on_save_loaded, SaveEventFn on_save_written, void* user_data,
-        SaveObserverHandle* out_handle);
+    ModResult (*observe_saves)(ModContext* ctx, SaveEventFn on_new_save, SaveEventFn on_save_loaded,
+        SaveEventFn on_save_written, void* user_data, SaveObserverHandle* out_handle);
 
-    /* Remove an observer previously registered by the calling mod. */
     ModResult (*unobserve_saves)(ModContext* ctx, SaveObserverHandle handle);
 
-    /*
-     * Read a named blob from any slot (0..2), including at file select when no slot is
-     * active. Same buffer contract as get_blob. The calling mod's own namespace only.
-     */
-    ModResult (*peek_blob)(ModContext* ctx, uint32_t slot, const char* name, void* buf, size_t* inout_size);
+    /* Read the calling mod's blob from any slot. Uses the get_blob buffer contract. */
+    ModResult (*peek_blob)(
+        ModContext* ctx, uint32_t slot, const char* name, void* buf, size_t* inout_size);
 
 } SaveService;
 
-MOD_DECLARE_SERVICE(
-    SaveService, svc_save, SAVE_SERVICE_ID, SAVE_SERVICE_MAJOR, SAVE_SERVICE_MINOR);
+MOD_DECLARE_SERVICE(SaveService, svc_save, SAVE_SERVICE_ID, SAVE_SERVICE_MAJOR, SAVE_SERVICE_MINOR);
