+++ src/dusk/utilities.hpp
@@ -2,7 +2,16 @@
 
 #include <cstddef>
 #include <cstdint>
+#include <string>
+#include <vector>
+
+#ifdef crc32
+// miniz defines crc32 as an alias.
+#undef crc32
+#endif
 
 namespace dusk::utils {
-uint32_t CRC32(const void* data, size_t size);
-}
+std::string base64_encode(const std::vector<uint8_t>& data);
+bool base64_decode(const std::string& text, std::vector<uint8_t>& out);
+uint32_t crc32(const void* data, size_t size);
+}  // namespace dusk::utils
